Add files via upload
# Amazon-Clone
Welcome to my Amazon Clone project repository! This dynamic e-commerce website is built with HTML, CSS, and JavaScript, showcasing a responsive design and seamless user interactions. Dive into the code to explore the intricacies of building a modern online shopping platform. Join me on this coding journey as we replicate the magic of Amazon in every line!